test(integrations): make ZIO Prelude tests pass on both Scala 2 and 3
